# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Automarketti API",
    "description": "Automarketti site API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": ""
    },
    {
      "url": "https://automarketbackend.onrender.com",
      "description": ""
    }
  ],
  "tags": [
    {
      "name": "Item",
      "description": "Individual announcement endpoints"
    },
    {
      "name": "Items",
      "description": "All announcements endpoints"
    },
    {
      "name": "Users",
      "description": "Users endpoints"
    },
    {
      "name": "User",
      "description": "User profile endpoints"
    }
  ],
  "paths": {
    "/api/items/": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Response all announcements",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/item/": {
      "post": {
        "tags": [
          "Item"
        ],
        "summary": "Save a new car announcement",
        "description": "",
        "responses": {
          "201": {
            "description": "Response back saved car, id and message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mark": {
                      "type": "string",
                      "example": "Volvo",
                      "description": "Car mark"
                    },
                    "model": {
                      "type": "string",
                      "example": "V70",
                      "description": "Car model"
                    },
                    "fuelType": {
                      "type": "string",
                      "example": "diesel",
                      "description": "Car fuel type"
                    },
                    "mileage": {
                      "type": "number",
                      "example": 120000,
                      "description": "Car mileage or odometer reading"
                    },
                    "gearBoxType": {
                      "type": "string",
                      "example": "manual",
                      "description": "Car gear box type"
                    },
                    "price": {
                      "type": "number",
                      "example": 10000,
                      "description": "Car price"
                    },
                    "description": {
                      "type": "string",
                      "example": "Hyvä auto",
                      "description": "Announcement description"
                    },
                    "photos": {
                      "description": "Car photos",
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "mark": {
                    "type": "string",
                    "example": "Volvo",
                    "description": "Car mark"
                  },
                  "model": {
                    "type": "string",
                    "example": "V70",
                    "description": "Car model"
                  },
                  "fuelType": {
                    "type": "string",
                    "example": "diesel",
                    "description": "Car fuel type"
                  },
                  "mileage": {
                    "type": "number",
                    "example": 120000,
                    "description": "Car mileage or odometer reading"
                  },
                  "gearBoxType": {
                    "type": "string",
                    "example": "manual",
                    "description": "Car gear box type"
                  },
                  "price": {
                    "type": "number",
                    "example": 10000,
                    "description": "Car price"
                  },
                  "description": {
                    "type": "string",
                    "example": "Hyvä auto",
                    "description": "Announcement description"
                  },
                  "photos": {
                    "description": "Car photos",
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "mark",
                  "model",
                  "fuelType",
                  "mileage",
                  "gearBoxType",
                  "price",
                  "photos"
                ]
              }
            }
          }
        }
      }
    },
    "/api/item/{id}": {
      "get": {
        "tags": [
          "Item"
        ],
        "summary": "Get individual car announcement",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Item"
        ],
        "summary": "Delete individual car announcement",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/photo/{fileName}": {
      "get": {
        "tags": [
          "Item",
          "Items"
        ],
        "summary": "Announcement photos download",
        "description": "",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/registration/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User registration",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Users login endpoint",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Responses user login details in JSONWebToken",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciplJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3RpIiwicm9sZeI6InVzZXIiLCJpZCI6xjY3NTAxYjMzODAxNjU4MDliMjUxMzRlNSIsImlhdCI6MTczMzMwMzcwMSwiZXhwIjoxNzMzMzA3MzAxfQ.aLJeXX_VINMwL1i_n9ad0OOJflObUTZHbzQiljt01Jw"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "All users profile data",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Single user profile data",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/message/": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Messages between users",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "example": "any"
                  },
                  "userId": {
                    "example": "any"
                  },
                  "recipientUserId": {
                    "example": "any"
                  },
                  "announcementId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "mark": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "model": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "fuelType": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "mileage": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "number"
              }
            }
          },
          "gearBoxType": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "price": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "number"
              }
            }
          },
          "description": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "photos": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "items": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "binary"
                  }
                }
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "password": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          }
        }
      }
    }
  }
}